chore: use nextest (#807)
* Use nextest in CI.
* Move one libp2p test to the integration test folder
