+++ Makefile
@@ -43,6 +43,10 @@ test: $(AUTOGENERATED_BLOCKLIST_CLIENT_CLIENT) $(AUTOGENERATED_EMILY_CLIENT)
 	cargo test -- --test-threads=1
 	pnpm --recursive test
 
+test-ci: $(AUTOGENERATED_BLOCKLIST_CLIENT_CLIENT) $(AUTOGENERATED_EMILY_CLIENT)
+	cargo nextest run --no-fail-fast 
+	pnpm --recursive test
+
 lint: $(AUTOGENERATED_BLOCKLIST_CLIENT_CLIENT) $(AUTOGENERATED_EMILY_CLIENT)
 	cargo clippy -- -D warnings
 	# TODO(719): No need to do this once rustfmt 2.0.0 ships
@@ -68,6 +72,9 @@ integration-env-up:
 
 integration-test: $(AUTOGENERATED_BLOCKLIST_CLIENT_CLIENT) $(AUTOGENERATED_EMILY_CLIENT)
 	cargo test --test integration --all-features --no-fail-fast -- --test-threads=1
+
+integration-test-ci: $(AUTOGENERATED_BLOCKLIST_CLIENT_CLIENT) $(AUTOGENERATED_EMILY_CLIENT)
+	cargo nextest run --test integration --all-features --no-fail-fast --test-threads=1
 
 integration-env-down:
 	docker compose --file docker/docker-compose.test.yml down -v
