+++ signer/tests/integration/main.rs
@@ -3,6 +3,7 @@ use std::sync::atomic::AtomicU16;
 mod bitcoin_client;
 mod bitcoin_rpc;
 mod block_observer;
+mod communication;
 mod complete_deposit;
 mod contracts;
 mod emily;
